python dash research/requirements.txt
# Web Framework
dash==2.14.2
flask==3.1.2
gunicorn==21.2.0

# Data & Visualization
plotly==5.24.1
pandas==2.2.0
openpyxl==3.10.10
xlrd==2.0.1

# UI Components
dash-bootstrap-components==1.6.0

# Utilities
python-dotenv==1.0.0
pytz==2024.1

# Development
black==24.1.1
flake8==7.0.0
